Refactor all code
